requirements.txt
anyio==4.4.0
asyncio==3.4.3
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.4.0
prettytable==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.3
python-telegram-bot-calendar==1.0.5
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.0.3
